+++ packages/flutter/lib/wenyousite_foundation.dart
@@ -2,4 +2,5 @@ library;
 
 export 'src/foundation_tokens.dart';
 export 'src/foundation_formatters.dart';
+export 'src/foundation_brand.dart';
 export 'src/wenyou_icons.dart';